refactor(Lotto): LottoErrorMessage 클래스 삭제 및 상수들을 Lotto로 이동
아무래도 Lotto관련 예외 메세지는 Lotto 클래스가 관리하는게 맞다고 판단하였습니다. 때문에, Lotto 클래스로 이동시키고, 기존의 LottoErrorMessage 클래스는 삭제했습니다.
