+++ src/test/java/lotto/domain/lotto_numbers/BonusNumberTest.java
@@ -1,6 +1,6 @@
 package lotto.domain.lotto_numbers;
 
-import static lotto.domain.lotto_numbers.LottoNumbersErrorMessages.INVALID_RANGE_OF_LOTTO_NUMBER;
+import static lotto.domain.lotto_numbers.Lotto.INVALID_RANGE_OF_LOTTO_NUMBER;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.params.ParameterizedTest;
